Где-то косяк

+++ core/handlers/privatechat_functions/eraser.py
@@ -4,6 +4,7 @@ file for eraser func
 from aiogram import types
 
 from core.database_functions.db_functions import delete_user
+from core.utils.is_chat_admin import is_chat_admin
 
 
 async def eraser(message: types.Message):
@@ -13,5 +14,10 @@ async def eraser(message: types.Message):
     :param message: admin send command to bot
     :return: sends message about success
     """
+    is_admin = await is_chat_admin(message.from_user.id)
+    if not is_admin:
+        await message.answer('Вы не являетесь модератором, функция недоступна')
+        return
+
     await delete_user(message.from_user.id)
     await message.answer('Успешно')
